Generate logout URL

Generates logout URL from configuration file. This can be used inside a shiny app button to logout from the app.

auth0_logout_url(config_file, redirect_js = TRUE)

Arguments

config_file
YAML file containing auth0 configuration.

redirect_js
include javascript code to redirect page? Defaults to TRUE.

Value

url string to logout, collapsed or not by javascript code.

Details

To use this function successfully inside a shiny app, you may want to install shinyjs package. See example

Examples

library(shiny)
library(auth0)
library(shinyjs)

#> 
#> Attaching package: ‘shinyjs’

#> The following object is masked from ‘package:shiny’:
#> 
#>     runExample

#> The following objects are masked from ‘package:methods’:
#> 
#>     removeClass, show

# configuration file generated by auth0::use_auth0()
config_file <- "_auth0.yml"

# simple UI with action button
# note that you must include shinyjs::useShinyjs() for this to work
ui <- fluidPage(shinyjs::useShinyjs(), actionButton("logout_auth0", "Logout"))

# server with one observer that logouts
server <- function(input, output, session) {
  observeEvent(input$logout_auth0, {
    # javascript code redirecting to correct url
    js <- auth0_logout_url(config_file)
    shinyjs::runjs(js)
  })
}

shinyAuth0App(ui, server, config_file = config_file)

#> Warning: cannot open file '_auth0.yml': No such file or directory

#> Error in file(file, "rt", encoding = fileEncoding): cannot open the connection